refactor(admin-ui):dialogs to separate folder
