+++ apps/sensenet/src/components/ContentListPanel.tsx
@@ -17,7 +17,7 @@ import {
 import { useRepository } from '../hooks'
 import { ContentBreadcrumbs } from './ContentBreadcrumbs'
 import { ContentContextMenu } from './ContentContextMenu'
-import { DeleteContentDialog } from './DeleteContentDialog'
+import { DeleteContentDialog } from './dialogs'
 import { DropFileArea } from './DropFileArea'
 import { Icon } from './Icon'
 import { SecondaryActionsMenu } from './SecondaryActionsMenu'
