+++ apps/sensenet/src/components/dialogs/index.ts
@@ -1,0 +1,5 @@
+export * from './add'
+export * from './content-info'
+export * from './copy-move'
+export * from './delete'
+export * from './edit-properties'
